face-recognition-attendance/backend/requirements.txt
# Web Framework
flask==2.3.2
flask-cors==4.0.0

# Authentication
flask-jwt-extended==4.5.2
flask-bcrypt==1.0.1

# Database
pymongo==4.3.3
dnspython==2.3.0

# Image Processing
opencv-python==4.7.0.72
face-recognition==1.3.0
numpy==1.24.3
Pillow==9.5.0

# Environment and Utilities
python-dotenv==1.0.0

# Optional but recommended
requests==2.30.0
logging==0.5.1.2

# Machine Learning and AI
cmake==3.26.3
dlib==19.24.1